backend/src/apps/recipes/serializers/category.py
from rest_framework import serializers

from src.apps.recipes.models.category import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ["id", "name"]
