backend/pyproject.toml
[tool.poetry]
name = "cookpedia"
version = "0.1.0"
description = "Home Recipe website"
authors = ["Szaneron <58951668+Szaneron@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
setuptools = "^74.1.2"
pyyaml = "^6.0.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
